testing/web-platform/tests/old-tests/submission/Infraware/Forms/contents/Forms/fieldset_form.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>FieldSet_form</h3>
  </p>
  
  <hr>

  <div id="log"></div>

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action="" 
      name="input_form"
      id="input_form">
    <fieldset id="input_field">
     </fieldset>
  </form>
  <script>

    var field = document.getElementById("input_field");

    if (typeof(field.form) == "object") {
      test(function() {
        assert_equals(field.form.name, "input_form", "form attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("form attribute is not exist.");
      });
    }

  </script>

 </body>
</html>
